transportation/requirements.txt
~umpy==1.19.0
opencv_python==4.4.0.46
thop==0.0.31.post2005241907
cupy_cuda100==7.5.0
matplotlib==3.2.2
torchsummary==1.5.1
torch==1.2.0
cupy==8.3.0
numpy==1.19.4
Pillow==8.0.1
